Move the banner function from versions.jl to REPL
